move implementation to redisutils
